Get font's loading
I've not managed to get fonts loading via the asset pipeline in both
production and development with the same configuration.  Currently,
ct-vis assets are not served by apache in production.  We also need to
prefix them with a `--/`.  I can't find a single mechanism to have
everthing work with the asset pipeline.

Instead, they are being served from the public directory in production.
This is unlikely to cause an issue as they are not expected to change.