Authorization via cancancan
Authorization has been added to most all controller actions.  There are
a few that it has not been added to as it is not immediately clear what
should be authorized.  Those actions don't have authorization in emma
either.

A consequence of adding authorization is that the context menu on the
IRV now shows a number of actions that do not work/have not been ported.
This needs dealing with at some point.