Create auth token when signing in via JSON request
The creation and revocation of tokens is delegated to `devise-jwt` /
`warden-jwt_auth`.  This ensures that each token is associated with a
particular user which makes authorisation simpler than having tokens
disconnected from users.

These tokens are sufficient for the rack/device API.  There is currently
some difficulty in getting them to work with the metric reporting API.

Moving forward it is likely that we will want to have more control over
the generated tokens.  This is possible by manual calls to
`Warden::JWTAuth::TokenEncoder` / `Warden::JWTAuth::UserEncoder`.