Remove support for reverse data source map
There is a ruby marshal versioning error between Ruby 3.0 and Ruby 1.9
despite them both claiming to use the same Marshal format.  Or perhaps
it is an issue with Dalli::Client?  Not sure.  However, it seems that
only Demm requires this so its good for us to ignore it for now.