Fix race condition when loading IRV
Under some circumstances, the `testLoadProgress` function was not called
when the non-rack device definitions were retrieved.  If the non-rack
device definitions were the last resource to be retrieved the IRV would
get stuck at "80%" loading. The solution is to ensure that
`testLoadProgress` is called under all circumstances.

Whilst debugging this, the `AssetManager` saw some love, and logging for
the `AssetManager` and `IRVController` was improved.