Fix issue with array methods added to presets
Previously, prior to the presets being retrieved, the method presetsById
was returning an array resulting in the preset dropdown being populated
with array method names instead of an empty list of preset names.

This commit fixes this issue by setting the default set of presets to an
object instead of an array.  It also modernises the iteration method.