Hacor migration migration
Migrate some database migrations from hacor to ivy engine.  The
following changes have been made to the domain model.

* The Cluster model has gone. If we want to group racks, we'll figure
  out the correct way when we have a use case, probably group by
  customer to begin with.
* The DataSource model has gone. DataSourceMap remains, but we no longer
  pretend to have support for something other than ganglia.
* Device role attributes along with mia/isla methods have gone.
* The unused Device subclasses have been removed.
* The created_on attribute is renamed to created_at.

The serialization of racks for the IRV needed some love following a
change to how templates serialize their images.

Whilst fixing that, the issue with the IRV being broken with 0 racks has
also been improved.  Instead of getting stuck at 80% loading, it now
fully loads and presents 0 racks.  An improvement would be to display a
message to the user informing them that there are no racks.