API to create, update and destroy templates
Only admin users can create, update or destroy templates.

By default, the API refuses to destroy templates have had devices
created from them. This can be overridden with a `recurse=true`
parameter. If this is given all devices created from the template are
destroyed along with the template.

A template's name and description can be updated, but not its height.
This prevents an issue where updating the templates height could result
in overlapping devices in a rack.

Example scripts have been added for the new API.

The template list API has been updated to include the template's height.