Add project ID field to non-admin users
Project ID is an opaque string of 255 characters.  The concertim
openstack service will use it for its processing.  Currently concertim,
itself, does not.

The project id can be set when an account is created.  It can also be
updated from the account details page.

The project id is inluded in the `list-users` and `show-user` responses.
It is also included in the owner section of the `list-racks` and
`show-rack` respones.