Include new metadata in IRV tooltip
The rack metadata is displayed in the rack tooltip after all other
tooltip data.  The order of the keys is ill-defined, but consistent. If
ordering is important we can look at changing this.

The device tooltip has been changed to have the new template fields
(VCPUs, RAM and disk) after the device's description.  Then the device
metadata fields in a consistent, but ill-defined, order.  Then the
template name and description.