Prototype for sending cloud env config to a server
There is a button in the cloud env config page actions dropdown to send
the configuration.  When clicked the cloud environment config is POSTed
as JSON to the IP and port given in the cloud environment config.

Currently, the response is ignored other than to test whether it was
successful or not.

If an error occurs, the user is informed of this and they have the
option of trying again.  As this is currently a fully manual process, no
automated retry handling has been written.

It should be relatively simple to configure the IP, port, path and body
of the request for different use-cases.