Inform user creation service of user signup
Extract common job runner class to `Emma::Faraday::JobRunner`. ActiveJob
classes that need to make HTTP requests to our other services are
expected to use a `Runner` that inherits from
`Emma::Faraday::JobRunner`.  The derived class can override the `url`
and `body` methods to configure the request; it can override the `call`
method to add post processing of the response.