Retry failed new user jobs 10 times over 7 hours
If the job fails due to a network connection error or due to the user
handler not correctly processing the job, it will be retried with an
exponentially longer backoff.  It will be attempted 10 times over the
course of about 7 to 8 hours.

Common errors are raised as Faraday::Errors to cause retrying too.