Sign up job updates user with project and user ids
If the request is successful, but the response is not valid, e.g., it
doesn't contain `user_id` or `project_id`, the job will be retried.

Concertim has so far been careful to avoid refering to openstack.  So
the user's openstack id is referred to as `cloud_user_id`.

The `cloud_user_id` is sent to the user handler service if it is
present.  I'm not sure that would ever be the case, but it could be.