Devices cannot be moved to another rack
This works around an issue with device name uniqueness and also avoids
awkward questions such as what does it mean for an openstack instance to
be moved to another project belonging to another user?  And what does it
mean for a device belonging to a web hosting cluster to be moved to a
HPC cluster?