Overlay build status on chassis/devices
A new option has been added to the metrics drop down to display the
build status.  When displaying the build status merics cannot also be
show and images are always shown.

The IRV controls need more thought. Perhaps we should have a "metric
mode", a "build status" mode and an "infrastructure mode".