Don't update openstack password
Currently, we only want the openstack password to be set once.  If the
user changes their password, their concertim password is to be updated,
but not their openstack password.

When the openstack user handler service supports updating the openstack
user's password we will remove this restriction and have Concertim
create a job to inform the user handler to update the password.