Add API to return data source maps
The format allows mapping from a data source map to the device's ID. The
"map to host" part of the data source map is the device's ID prefixed
with the string `device:`.  This is used in the name of the RRD file
created by ganglia and is not so easy to migrate, so we include the
`device:` prefix in case it is ever needed.

Metric daemon will construct the reverse lookup if it needs to do so.