Move importmap JavaScript files around
All of the JavaScript files moved in this commit are imported using
`config/importmap.rb`.  The path used to import them has not changed
allowing all files to be moved without any changes to them being
required.

If the path given in `config/importmap.rb` doesn't match the new path
the following error would appear in the javascript console.  As it does
not it can be taken as a sign that the changes are consistent and
nothing has been broken by this move.

```
Uncaught TypeError: The specifier “canvas/irv/util/Parser” was a bare specifier, but was not remapped to anything. Relative module specifiers must start with “./”, “../” or “/”.
```

The aim is to remove the `mia` and `coffee` directories to avoid
questions such as "what's a mia?" and "what's coffee got to do with
this?".

* All files and directories under `app/javascript/mia/coffee/src/` have
  had the `mia/coffee/src/` path component removed.
* The files at `app/javascript/mia/javascript/irv/` have moved under
  `app/javascript/canvas/irv/`.
* The files at `app/javascript/mia/javascript/util/` have moved under
  `app/javascript/util/`.
* The `config/importmap.rb` has been updated accordingly.

This leaves the files at `app/javascript/mia/contrib/` which still need
rehousing.