Move mootools to vendor/javascript
If the changes were not consistent the following error would appear when
loading the interactive rack view.

```
Sprockets::FileNotFound in InteractiveRackViews#show
```

If mootools were not included, the following message would appear in the
javascript console.

```
Uncaught ReferenceError: Class is not defined
```

As neither error occurs t the changes can be considered consistent and
to have not broken anything.