Fix race condition loading IRV
The tests for whether the IRV had loaded did not include checking
whether the rule based access controls (RBAC) resource had loaded.

If the race condition was lost, the IRV would break following a
TypeError exception being raised whilst checking a user's permissions.

This commit fixes that race condition by including the RBAC in the set
of resources that we are waiting for.