Navigation item to show invoice
* Add a "view invoice" item to the navigation header under the account
  menu.
* Clicking the link opens a new tab showing the user's invoice.
* A request is made to the user concertim openstack hanlder to get the
  invoice.  That request contains the user's `billing_acct_id`, but not
  the current billing period.
* If successful, the invoice is rendered as is in the new tab.  We can
  later consider if it should be rendered "within the concertim layout".

The implementation uses a standard rails controller that runs an active
job to retrieve the invoice.  An alternative would be to use
`Rack::Proxy` to proxy the request.  It should be possible to add in the
needed credentials as part of proxying the request.  However, that route
was not taken as there are some circumstances under which we don't want
to proxy the request.  These being when the admin user has not
configured the cloud environment or the user's account does not have a
`billing_acct_id`.  It may be possible to perform some conditional
proxying in a `Rack::Proxy` controller, by that option was not taken.