Changes to providing JWT secret
The JWT secret can be read from an env variable or from a file given in
an env variable.  It is also now possible to have the app start without
a JWT secret.  This will obviously fail if we attempt to use JWT stuff,
but will allow some rake tasks to be ran such as precompiling assets.