Remove unused reference to default_db_table
This was used in legacy to allow sorting on columns that are from
another database table.  E.g., a table of devices might include details
taken from the device's template.  If so, sorting on them would require
the database table name in addition to the database column name.  We
currently, don't use that, so I've removed it.