Add invoice listing page
Display a resource table of invoices.  Each invoice row contains a
dropdown actions menu.  Eventually, this will include a link to show the
invoice.  Currently it links to the draft invoice.

The page also contains a link to show the draft invoice in the top
right-hand corner "actions" menu.

The ID generation for the per-row dropdown actions menu in the resource
table has changed.  Previously, it used the record's `id` attribute
(unless the record responded to `to_gid_param`).  However, invoices do
not have a `id` attribute, instead they have an `invoice_id` attribute.
The ID generation now uses `model_name.param_key` and `to_param`, which
is compatible with any `ActiveModel::Model` and should be sufficient to
produce unique ids in all cases.

In order for the invoices to support that mechanism, they have gained
both a `to_key` and `persisted?` method.  These methods also allow for
correctly creating links for invoices.