Paginate invoices
The default implementation of Pagy requires that the count of items is
available prior to the list of items being retrieved. Unfortunately, we
don't have easy access to the number of invoices until after we've
queried the API.

The number of invoices could be obtained by requesting 0 invoices with
an offset of 0, however, I don't like the idea of making multiple
requests to the billing middleware each time we load the invoices page.

Instead I have implemented a custom `Pagy` class `Pagy::DelayedCount`
similar to the `Pagy::Countless` class that Pagy ships with.  With this
class we can (1) use Pagy to calculate the page and items per page; (2)
make the request; (3) finalize the pagy instance with the count; (4) use
the pagy instance to render the pagination controls.  This provides
consistent handling of request parameters, defaults, variable errors and
pagination controls.