Fix issues with user sign up
The response from the user handler has renamed the key `billing_acct_id`
to `billing_account_id`.  This commit makes the corresponding change to
the response handler.

The repsonse handling code has changed to allow syncing the response in
two phases.  First the data relevant to the cloud environment is
validated and synced (the cloud_user_id and project_id).  If that is
valid, the data relevant to the billing platform is validated and synced
(the billing_acct_id).  If either of them fail, the error is now logged
and the job is requeued.

Also, I've increased the default timeout for making the HTTP request
from 5 seconds to 60 seconds.