Update invoice item format
The middleware sends invoice items in a new format.  It contains details
extracted from the killbill subscription, and only contains items that
we want to display.

Eventually, we are likely to want to provide a link from the invoice
item to the rack that it references.  We don't have a suitable target to
link to at the moment though.