Users can be deleted
When a request to delete a user is made, an asynchronous job is
scheduled to request the middleware to delete both the openstack and
billing accounts.  If that request is succesful, the concertim user is
deleted.  If not, the job is reattempted with an expoential backoff.

When a request to delete a user is made any existing sessions are
invalidated, it is no longer possible to sign in as that user and any
existing JWTs for that user are invalidated.

The user listing page has been updated to include the status of the
user; either "Active" or "Pending deletion".  The actions dropdown, also
includes an option to delete the user.