Enable browser validations
Browser validations for `required`, `min` and `max` have been enabled.

Validations for `pattern`, `minlength` and `maxlength` can be opted into
on a per-input basis.

Enabling `pattern` automatically doesn't really work.  The regex
needs to be anchored to the start and end of the string.  Rails
complains if `^`and `$` are used for this, instead wanting `\A` and
`\Z`.  JavaScript doesn't recognize those assertions.  That difference
of regular expression interpretation makes its automatic inclusion
awkward.  We'd also, need to specify the pattern in the help or hint
text too.

I don't like the behaviour of the `maxlength` browser validation.  It
silently prevents additional characters from being entered.  This is
easy to miss especially when pasting.  A better UX is to allow the input
but then complain on form submission.  The roundtrip validation provides
us with that behaviour.

The `minlength` validation is left as opt-in for parity with
`maxlength`.  Perhaps, it ought to be enabled automatically.

Missing validations from the user model have been added to make it
consistent with the database schema.