Migrate sequential integer primary keys to UUIDs
All of visualiser's models now use a uuid for their primary key.  To
avoid awkward data migrations, as much data as possible has been
deleted.  This is fine as we haven't had a production release yet.

The metrics being returned from the metric reporting daemon were
previously using string encoded integers for the IDs.  This has required
the removal of a `.to_i` call on the returned metrics body.

The IRV frequently used string encoded integers for the rack, chassis,
slot and device IDs.  I think this was because the racks were being
loaded for the IRV from an XML generating Postrgres query. The XML lost
the data type so the rack et al IDs were being sent to the browser as
string encoded ints.  The IRV worked around that by adhoc parsing of the
IDs to ints.  This has now been removed.