Send billing account id in create cluster request
The cluster builder is going to check that the user has sufficient
credits and create a killbill subscription when a cluster is created. To
do this it needs the user's billing account ID.

It also needs to know the URL at which the middleware service can be
reached.  We send it in the request too.