Inform middleware when user changes their password
When a user changes their password, a copy of it is stored in a new
encyrpted column `pending_foreign_password`.  A new `UserUpdateJob` is
performed in the background.  When that successfully completes, the
users `foreign_password` is set to their `pending_foreign_password` and
their `pending_foreign_password` is cleared.

This allows the password to be kept in (eventual) sync between concertim
and the cloud whilst minimising the period during which the passwords
are out of sync.