Allow for more frequent metric polling on IRV
The minimum metric refersh rate is now configurable via a new `Setting`
class.  The `Setting` class is intended to become a collection of
whatever settings we wish to allow admin users to configure via the UI.
Currently, there is a single `metric_refresh_interval` setting.

That setting is used as both the minimum and default metric refresh rate
on the IRV page.  A migration creates an entry of 15 seconds, which
matches the intervale of both the metric collection daemon and the
metric reporting daemon.

Future work will likely see a mechanism to allow the metrics to be
pushed from the metric reporting daemon to the IRV. For now we have a
mechanism to increase the refresh rate whilst still being able to
throttle it globally should 15 seconds prove to be too aggressive.