Extract drag handler from RackSpace
RackSpace is too large with too many responsibilities.  This commit
extract the drag handling to its own class.  It also applies the
decaffeinate suggestions and some other minor code tidy.

The structure of RackSpaceDragHandler is essentially the same as it was
when the code lived in RackSpace.  It is easier however to see which
code is related to dragging and to improve on it from here.