Add dedicated rack page
It contains an overview tab listing the contents that was previously in
the IRV rack hover tooltip, and a devices tab that contains a resource
table of devices belonging to the rack. Each device row links to the
device's details page.

Also add a spacer to IRV context menu for devices and racks to separate
destructive from non-destructive actions.