Fix issue with hint and chassis level metrics
The check for metric level was previously inside of RackSpace, where
`this.metricLevel` existed.  It was also kept in sync with the view
model's metricLevel.  RackSpaceHinter does not have access to
RackSpace's metricLevel, but can use the view model's metricLevel
directly.

Not that we actually have chassis level metrics at the moment though,
and perhaps we will never have them again.