Render certain constraints as drop downs
Some custom constraints such as `glance.image` can be rendered as a drop
down from which the user can select the desired image.  This commit adds
support for such dropdowns for `glance.image`, `neutron.network`,
`nova.flavor` and `nova.keypair`.

If the field also has an `allowed_values` constraint, those are used in
preference to the cloud assets.

The assets potentially differ for each user so they are retrieved from
cluster builder on each request.  Currently no attempt is made to cache
the values; we need cluster builder to be running anyway.