Move `Device`'s compute-specific fields to their own model.
A `Device` must have an associated `Device::ComputeDetails` model, which
contains those fields formerly in `Device` that are specific to compute
devices.

We render these out as an inline `details` object in the API, with a
`type` attribute to tell us which sort of details thing it is - though
at the moment we only have one sort of details thing. This is probably
going to change, but it gets most tests passing except for those which
are creating or updating a `Device`.