Continue to support partial updates.
If not all of the parameters of a `Device` are specified in a `PATCH`
request, then only the specified parameters are updated. This commit
makes that explicit by having tests that exercise that functionality,
and allows `details` to be absent or partial.

I'm not sure if it makes sense to support changing the type of the
related `details` object, since a `Device` is unlikely to change from
a compute to a network device etc., but I'm leaving that in for now.