Modify API to maintain support for current middleware.
We want the current middleware to keep working (we can't currently make
changes to it as it's in the middle of a big rewrite/refactor). This
commit lets the controllers accept both old and new data structures, and
the API return the old format (with compute attributes inlined into
`device` rather than being in a `details` object).

Once the middleware can be updated, this commit can be reverted.