Cluster launch and rack pages contain instructions
The cluster launch page displays the launch instructions for the cluster
type if any.

For all non-launch instructions, the rack page contains a tab for those
instructions.  Effectively, this results in a new "Usage instructions"
tab and possibly a new "Admin instructions" tab.

Displaying the correct instructions on the rack page requires visualiser
to be able to map from rack to cluster type.  We achive this by having
each stack's ouput contain a `concertim_cluster_type` value.  Visualiser
uses that to locate the cluster type that was used to launch the cluster
and pull the admin and usage instructions from it.

A better approach would be to have Visualiser create a Cluster model
when the launch request is made.  Bookkeeping such as the cluster type
used could then be added directly to that cluster model.  However, given
the discussion around concertim.com, I suspect that it is best to wait
until that development is underway.

Also, if the stack ouputs contain a `web_access` output, we use that to
link to the cluster's web access.  Presumably, this is Flight Web Suite.